+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `sla_domain` Terraform module.
 * Rename `aws_cnp_account` to `aws_iam_account` and move it to the `modules/` directory.
 * Update the `aws_exocompute` tests to use the `aws_iam_account` module.
+* Move the `aws_exocompute` module to the `modules/` directory.
 * Add support for the `KUBERNETES_PROTECTION` feature to the `aws_iam_account` module.
 * Add `gcp_exocompute` Terraform  module.
 * Add support for shared Exocompute to the `aws_iam_account` module.